backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.31.0
alembic==1.15.2
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.28.1
anthropic==0.52.0
twilio==9.6.0
cryptography==44.0.3
apscheduler==3.11.0
structlog==25.4.0
slowapi==0.1.9
